requirements.txt
langchain
langchain-openai
streamlit
python-dotenv
flask
twilio
pydantic
typing
